Some progress

+++ src/PlaywrightSharp/PointAndScroll.cs
@@ -4,9 +4,9 @@ namespace PlaywrightSharp
 {
     internal class PointAndScroll
     {
-        public decimal? ScrollX { get; set; }
+        public double? ScrollX { get; set; }
 
-        public decimal? ScrollY { get; set; }
+        public double? ScrollY { get; set; }
 
         public Point Point { get; set; }
     }
